encode jpeg2000

+++ dcmtk-3.5.4/dcmjpeg/include/dcmtk/dcmjpeg/djeijg2k.h
@@ -43,17 +43,17 @@ public:
 	Uint8 pixelRepresentation,
 	double minUsed, double maxUsed);
 
-virtual OFCondition encode( 
-  Uint16 columns,
-  Uint16 rows,
-  EP_Interpretation colorSpace,
-  Uint16 samplesPerPixel,
-  Uint8 * image_buffer,
-  Uint8 * & to,
-  Uint32 & length,
-  Uint8 bitsAllocated,
-  Uint8 pixelRepresentation,
-  double minUsed, double maxUsed);
+  virtual OFCondition encode( 
+    Uint16 columns,
+    Uint16 rows,
+    EP_Interpretation colorSpace,
+    Uint16 samplesPerPixel,
+    Uint16 * image_buffer,
+    Uint16 * & to,
+    Uint32 & length,
+    Uint8 bitsAllocated,
+    Uint8 pixelRepresentation,
+    double minUsed, double maxUsed);
   
   
   /** returns the number of bytes per sample that will be expected when encoding.
